Rename standalone package to hbaseendpoint
